feat: move state into root drawer component
