+++ apps/studio/src/features/editing-experience/components/EditPageDrawer.tsx
@@ -11,11 +11,6 @@ import { BsPlus } from 'react-icons/bs'
 import { Button } from '@opengovsg/design-system-react'
 import RootStateDrawer from './RootStateDrawer'
 
-type Block = {
-  text: string
-  id: string
-}
-
 export type RootDrawerState = {
   state: 'root'
   blocks: Block[]
@@ -47,26 +42,9 @@ type EditPageDrawerProps = {
 export const EditPageDrawer = ({ open, state }: EditPageDrawerProps) => {
   const [currState, setCurrState] = useState<DrawerState>(state)
 
-  const onDragEnd = (result: DropResult) => {
-    if (!result.destination) return
-
-    if (currState.state === 'root') {
-      const updatedBlocks = Array.from(currState.blocks)
-      // Remove block at source index
-      const [movedBlock] = updatedBlocks.splice(result.source.index, 1)
-      // Insert at destination index
-      updatedBlocks.splice(result.destination.index, 0, movedBlock!)
-
-      setCurrState({
-        ...currState,
-        blocks: updatedBlocks,
-      })
-    }
-  }
-
   switch (currState.state) {
     case 'root':
-      return <RootStateDrawer onDragEnd={onDragEnd} currState={currState} />
+      return <RootStateDrawer blocks={currState.blocks} />
     default:
       return <h1>Edit Page Drawer</h1>
   }
